Module Sample.Int

val range : int -> int -> int t

range min max is a sample that produces a value x such that min <= x < max.

val small : int t

small is a sample producing small integers in the rante -10 to 10.

val medium : int t

medium is a sample that produces value in the range -1000 to 1000.

val positive : int t

positive is a sample that produces positive values.

val negative : int t

negative is a sample that produces negative values.